feat: entity type selector on search page
